# docs/engineering/openapi.yaml
openapi: 3.1.0
info:
  title: CreatorHub API
  version: 0.1.0
  description: |
    CreatorHub API — travel social platform + experience marketplace for India.
    All monetary amounts are in **paisa** (integer). All timestamps are ISO 8601 UTC.
    Auth: Firebase Phone OTP → JWT verified server-side.
  contact:
    name: CreatorHub Engineering
  license:
    name: Proprietary
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://api.creatorhub.in
    description: Production

# ─── SECURITY ────────────────────────────────────────────────
security:
  - bearerAuth: []

# ─── TAGS ────────────────────────────────────────────────────
...
paths:
...
  /api/v1/onboarding/verticals:
    put:
      operationId: setOnboardingVerticals
      tags: [Onboarding]
      summary: Set user's vertical interests (minimum 3)
      description: |
        Writes to `user_active_verticals`. Verticals with ≤10 published creators
        also written to `user_waitlisted_verticals`.
        SRS: ONB-FR-003, ONB-FR-008
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [verticals]
              properties:
                verticals:
                  type: array
                  items:
                    $ref: "#/components/schemas/Vertical"
                  minItems: 3
                  maxItems: 8
      responses:
        "200":
          description: Verticals set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse-VerticalsResult"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/v1/verticals:
    get:
      operationId: listVerticals
      tags: [Onboarding]
      summary: List all verticals with creator counts
      description: |
        Returns all 8 verticals with live creator counts.
        Verticals with 0 creators marked "coming soon".
        SRS: ONB-FR-007
      security:
        - {}
        - bearerAuth: []
      responses:
        "200":
          description: Verticals with counts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse-VerticalsList"
  /api/v1/cities:
    get:
      operationId: searchCities
      tags: [Onboarding]
      summary: Search cities (~4000 Indian cities)
      description: |
        Fuzzy search across ~4000 cities. Used in onboarding location picker.
        No tier filtering — all cities equally valid.
        SRS: ONB-FR-010
      security:
        - {}
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 1
          description: Search query (fuzzy matched against city name)
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
            maximum: 50
      responses:
        "200":
          description: Matching cities
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse-CitiesList"
...
  /api/v1/content:
    post:
      operationId: createContentDraft
      tags: [Content]
      summary: Create a new content draft
      description: |
        Creates a draft content record of the specified type.
        User must be authenticated and have completed onboarding.
        Returns the draft with its ID for subsequent updates.
        SRS: CRT-FR-007, CRT-FR-008, CRT-FR-015
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContentDraftRequest"
      responses:
        "201":
          description: Draft created
          headers:
            Location:
              schema:
                type: string
              description: URI of created content resource
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse-ContentSummary"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          description: Onboarding not completed
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "429":
          $ref: "#/components/responses/RateLimited"
...
  /api/v1/posts:
    post:
      operationId: createPostDraft
      tags: [Posts]
      summary: Create a post draft
      description: |
        Creates a new post draft. Sets type='post', pricing_model='free', price_paisa=0.
        No KYC required for posts.
        SRS: CRT-FR-001
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostRequest"
      responses:
        "201":
          description: Post draft created
          headers:
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse-ContentSummary"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"

    get:
      operationId: listPosts
      tags: [Posts]
      summary: List published posts
      description: |
        Returns published posts. Supports filtering by vertical, creator.
        Cursor-based pagination. Guests allowed.
        SRS: CRT-FR-001
      security:
        - {}
        - bearerAuth: []
      parameters:
        - name: vertical
          in: query
          schema:
            $ref: "#/components/schemas/Vertical"
        - name: user_id
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by creator
        - $ref: "#/components/parameters/cursor"
        - $ref: "#/components/parameters/limit"
      responses:
        "200":
          description: Post list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse-ContentList"
...
  /api/v1/itineraries:
    post:
      operationId: createItineraryDraft
      tags: [Itineraries]
      summary: Create an itinerary draft
      description: |
        Creates a new self-paced itinerary draft with initial day skeletons.
        Requires day_count (1-30). Sets type='self_paced_itinerary'.
        SRS: CRT-FR-002, CRT-FR-019
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateItineraryRequest"
      responses:
        "201":
          description: Itinerary draft created with day skeletons
          headers:
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse-ItinerarySummary"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Firebase ID token or app session token

  # ── Parameters ──────────────────────────────────────────────
  parameters:
...
    cursor:
      name: cursor
      in: query
      schema:
        type: string
      description: Opaque cursor for pagination (base64-encoded)
    limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 50

  # ── Reusable Responses ──────────────────────────────────────
  responses:
    Unauthorized:
      description: Authentication required or token invalid/expired
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
    ValidationError:
      description: Request validation failed (Zod errors)
      content:
        application/problem+json:
          schema:
            allOf:
              - $ref: "#/components/schemas/ProblemDetail"
              - type: object
                properties:
                  errors:
                    type: array
                    items:
                      $ref: "#/components/schemas/FieldError"
...
    RateLimited:
      description: Rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds until rate limit resets
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"

  # ── Schemas ─────────────────────────────────────────────────
  schemas:
...
    Vertical:
      type: string
      enum:
        - travel
        - stories
        - food
        - fitness
        - education
        - photography
        - music
        - wellness
    ContentType:
      type: string
      enum:
        - post
        - event
        - scheduled_experience
        - self_paced_itinerary
    ContentStatus:
      type: string
      enum:
        - draft
        - under_review
        - published
        - unpublished
        - archived
        - rejected
        - taken_down
    Visibility:
      type: string
      enum: [public, unlisted, private]
    PricingModel:
      type: string
      enum: [free, paid]
...
    ProblemDetail:
      type: object
      description: RFC 9457 Problem Details
      required: [success, error]
      properties:
        success:
          type: boolean
          const: false
        error:
          type: object
          required: [type, title, status, detail, instance]
          properties:
            type:
              type: string
              format: uri
              example: https://creatorhub.in/errors/validation-failed
            title:
              type: string
              example: Validation Failed
            status:
              type: integer
              example: 400
            detail:
              type: string
              example: Title must be between 3 and 200 characters.
            instance:
              type: string
              example: /api/v1/itineraries
            errors:
              type: array
              items:
                $ref: "#/components/schemas/FieldError"
    FieldError:
      type: object
      required: [field, message, code]
      properties:
        field:
          type: string
          example: title
        message:
          type: string
          example: String must contain at least 3 character(s)
        code:
          type: string
          example: too_small
    CursorMeta:
      type: object
      properties:
        next_cursor:
          type: string
          nullable: true
          description: Opaque cursor for next page (null if no more results)
        has_more:
          type: boolean
        per_page:
          type: integer

    # ── Auth ──────────────────────────────────────────────────
...
    VerticalsResult:
      type: object
      required: [active_verticals, waitlisted_verticals]
      properties:
        active_verticals:
          type: array
          items:
            $ref: "#/components/schemas/Vertical"
        waitlisted_verticals:
          type: array
          items:
            $ref: "#/components/schemas/Vertical"
          description: Verticals with ≤10 published creators (auto-detected)
    VerticalInfo:
      type: object
      required: [slug, name, creator_count, status]
      properties:
        slug:
          $ref: "#/components/schemas/Vertical"
        name:
          type: string
          example: Travel
        creator_count:
          type: integer
          description: Live count of published creators in this vertical
        status:
          type: string
          enum: [active, soon, coming_soon]
          description: "active (>10 creators), soon (1-10), coming_soon (0)"
...
    City:
      type: object
      required: [id, name, state, country]
      properties:
        id:
          type: string
          example: in.mh.pune
        name:
          type: string
          example: Pune
        state:
          type: string
          example: Maharashtra
        country:
          type: string
          example: IN
        lat:
          type: number
          format: double
        lng:
          type: number
          format: double

    # ── Upload ────────────────────────────────────────────────
...
    CreateContentDraftRequest:
      type: object
      required: [type, vertical]
      properties:
        type:
          $ref: "#/components/schemas/ContentType"
        vertical:
          $ref: "#/components/schemas/Vertical"
        title:
          type: string
          minLength: 1
          maxLength: 100
          description: Optional at draft creation
        description:
          type: string
          maxLength: 280
...
    ContentSummary:
      type: object
      required: [id, type, vertical, status, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        type:
          $ref: "#/components/schemas/ContentType"
        vertical:
          $ref: "#/components/schemas/Vertical"
        title:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/ContentStatus"
        visibility:
          $ref: "#/components/schemas/Visibility"
        pricing_model:
          $ref: "#/components/schemas/PricingModel"
        price_paisa:
          type: integer
        cover_image_url:
          type: string
          format: uri
          nullable: true
          description: URL of the first media item (cover image)
        like_count:
          type: integer
        comment_count:
          type: integer
        view_count:
          type: integer
        published_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    ContentListResponse:
      type: object
      required: [items, meta]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ContentSummary"
        meta:
          $ref: "#/components/schemas/CursorMeta"
...
    CreatePostRequest:
      type: object
      required: [vertical]
      properties:
        vertical:
          $ref: "#/components/schemas/Vertical"
        title:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          maxLength: 280
...
    CreateItineraryRequest:
      type: object
      required: [vertical, day_count]
      properties:
        vertical:
          type: string
          enum: [travel]
          description: Itineraries are travel-only in MVP
        title:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          maxLength: 280
        day_count:
          type: integer
          minimum: 1
          maximum: 30
          description: Number of days — creates empty day skeletons
        starting_city_id:
          type: string
          description: Starting city for the itinerary
        destination_city_ids:
          type: array
          items:
            type: string
          maxItems: 10
        sub_category_id:
          type: string
          description: Travel sub-category (e.g., "travel.road_trips")
...
    ItinerarySummary:
      allOf:
        - $ref: "#/components/schemas/ContentSummary"
        - type: object
          properties:
            day_count:
              type: integer
            total_spots:
              type: integer
            total_distance_km:
              type: number
              format: float
              nullable: true
            starting_city:
              $ref: "#/components/schemas/City"
...
    SuccessResponse-VerticalsResult:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          const: true
        data:
          $ref: "#/components/schemas/VerticalsResult"
    SuccessResponse-VerticalsList:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          const: true
        data:
          type: array
          items:
            $ref: "#/components/schemas/VerticalInfo"
...
    SuccessResponse-CitiesList:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          const: true
        data:
          type: array
          items:
            $ref: "#/components/schemas/City"
...
    SuccessResponse-ContentSummary:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          const: true
        data:
          $ref: "#/components/schemas/ContentSummary"
...
    SuccessResponse-ContentList:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          const: true
        data:
          $ref: "#/components/schemas/ContentListResponse"
...
    SuccessResponse-ItinerarySummary:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          const: true
        data:
          $ref: "#/components/schemas/ItinerarySummary"
...